Fix the frontends to use the new api
did not fix rpi baremetal as it does not compile on the newer toolchains :(
