+++ Makefile.toml
@@ -62,7 +62,6 @@ args = ["component", "add", "rust-src"]
 [tasks.libretro]
 command = "cargo"
 args = ["build", "--release", "--package", "magenboy_libretro"]
-dependencies = ["libretro_android"]
 
 [tasks.libretro_android]
 install_crate = {crate_name = "cargo-ndk", binary = "cargo", test_arg = "ndk"}
