+++ common/Cargo.toml
@@ -16,6 +16,7 @@ chrono = {version = "0.4", optional = true}
 
 [features]
 std = ["chrono", "fern", "crossbeam-channel"]
+dbg = ["std"]
 
 [dev-dependencies]
 criterion = "0.3"
